# restraunat/restaurant-backend/app/Models/Category.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;

class Category extends Model
{
    protected $fillable = ['name', 'image', 'slug'];

    public function products()
    {
        return $this->hasMany(Product::class);
    }
}