# restraunat/restaurant-backend/app/Models/News.php
<?php namespace App\Models;
use Illuminate\Database\Eloquent\Model;

class News extends Model {
    protected $fillable = ['title','body','image','published_at'];
    protected $casts    = ['published_at' => 'datetime'];
}